+++ app/apk/src/main/java/oi/masksu/com/ui/home/HomeMaskSuContent.kt
@@ -45,7 +45,7 @@ internal fun LazyListScope.masksuHomeContent(
                     )
                     if (Info.env.isActive) {
                         UninstallButton(
-                            text = context.getString(CoreR.string.home_uninstall_weavemask),
+                            text = context.getString(CoreR.string.home_uninstall_masksu),
                             onPressed = { viewModel.onDeletePressed() }
                         )
                     }